When I load, add or try to delete agratitude it crashed cos the components are not yet accessible. I needed to add the authentication token to the files that needed them. In the Home.js comp., I added the header and called fetch gratitudes only when there is a user. In GratitudeDetails.js and GratitudesFooter.js, if there is no user, I don't even make the reqest. In GratitudeForm, I added the authorization and setError when there is no user and force it not to send the request. Solved!

+++ frontend/src/components/GratitudeDetails.js
@@ -1,11 +1,21 @@
 import { FaTrash } from 'react-icons/fa';
 import { useGratitudeContext } from '../hooks/useGratitudesContext';
+import { useUserContext } from '../hooks/useUserContext';
 
 const GratitudeDetails = ({ gratitudeItem }) => {
   const { dispatch } = useGratitudeContext();
+  const { user } = useUserContext();
+
   const handleClick = async () => {
+    // if there's no user, we dont even try to make request.
+    if (!user) {
+      return;
+    }
     const response = await fetch('/api/gratitudes/' + gratitudeItem._id, {
       method: 'DELETE',
+      headers: {
+        Authorization: `Bearer ${user.token}`,
+      },
     });
     const json = await response.json();
